@@ -19,9 +19,9 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>11</java.version>
-		<maven.compiler.source>${java.version}</maven.compiler.source>
-		<maven.compiler.target>${java.version}</maven.compiler.target>
+		<java.version>17</java.version>
+<!--		<maven.compiler.source>${java.version}</maven.compiler.source>-->
+<!--		<maven.compiler.target>${java.version}</maven.compiler.target>-->
 		<start-class>br.unb.scrap.Application</start-class>
 	</properties>
 
@@ -71,11 +71,13 @@
 			<scope>test</scope>
 		</dependency>
 
+		<!-- https://mvnrepository.com/artifact/org.springframework/spring-web -->
 		<dependency>
-			<groupId>org.apache.logging.log4j</groupId>
-			<artifactId>log4j-core</artifactId>
-			<version>2.14.1</version>
+			<groupId>org.springframework</groupId>
+			<artifactId>spring-web</artifactId>
+			<version>6.1.3</version>
 		</dependency>
+
 
 	</dependencies>
 
