@@ -1,23 +1,34 @@
 package br.unb.scrap.components;
 
 import java.io.IOException;
+import java.util.stream.Stream;
 
 import org.jsoup.nodes.Document;
 import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+
+import br.unb.scrap.utils.UrlUtils;
 
 public class ConnectionJsoupTest {
 
-	String url = ScrapBoost.BASE_URL;
-	ConnectionJsoup cj = new ConnectionJsoup();
+	String urlOpenJdk = UrlUtils.BASE_URL_OPENJDK;
+	String urlBoost = UrlUtils.BASE_URL_BOOST;
 
-	@Test
-	public void testJsoupConnection() {
+	private ConnectionJsoup cj = new ConnectionJsoup();
+
+	@ParameterizedTest
+	@MethodSource("provideUrls")
+	public void testJsoupConnection(String url) {
 		try {
 			Document document = cj.connect(url);
 			Assert.assertNotNull(document);
 		} catch (IOException e) {
 			Assert.fail("Falha na conexão: " + e.getMessage());
 		}
+	}
+
+	private static Stream<String> provideUrls() {
+		return Stream.of(UrlUtils.BASE_URL_OPENJDK, UrlUtils.BASE_URL_BOOST);
 	}
 }
