@@ -1,6 +1,6 @@
 package br.unb.scrap.components;
 
-import static br.unb.scrap.utils.UrlUtils.BASE_URL_BOOST;
+import static br.unb.scrap.utils.UrlUtils.BASE_URL_OPENJDK;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -8,9 +8,6 @@ import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -19,23 +16,22 @@ import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
 import br.unb.scrap.enums.PostTypeEnum;
 import br.unb.scrap.model.Post;
 
 @Component
-@Primary // Indica a implementação principal de um componente quando há várias opções para uma interface ou classe
-public class ScrapBoost implements PageScraper {
+public class ScrapOpenJDK implements PageScraper {
 
-	private static final Logger logger = LogManager.getLogger(ScrapBoost.class);
-	private static final String DATE_TAG_SELECTOR = "tbody > tr > td:nth-child(3)";
-	private static final String THREAD_TAG_SELECTOR = "tbody > tr > td:nth-child(4)";
-	private static final String DATE_URL_SUFFIX = "date.php";
+	private static final Logger logger = LogManager.getLogger(ScrapOpenJDK.class);
+	private static final String TR_TAG_SELECTOR = "tbody > tr";
+	private static final String DATE_TAG_SELECTOR = "td:nth-child(2)";
+	private static final String THREAD_TAG_SELECTOR = "tbody > tr > td:nth-child(1)";
+	private static final String DATE_URL_SUFFIX = "date.html";
 	private final FileLogger fileLogger;
 
-	public ScrapBoost() {
+	public ScrapOpenJDK() {
 		fileLogger = new FileLogger();
 	}
 
@@ -63,24 +59,30 @@ public class ScrapBoost implements PageScraper {
 
 	/**
 	 * método é responsável por obter os links dos posts por data a partir de um
-	 * documento HTML, a partir da BASE_URL_BOOST ou da url passada.
+	 * documento HTML, a partir da BASE_URL_OPENJDK ou da url passada.
 	 * 
 	 * @return Retorna a lista urls contendo as URLs das threads extraídas.
 	 */
 	public List<String> getLinksByDate() throws IOException {
 		List<String> dateUrls = new LinkedList<>();
 		try {
-			Document doc = Jsoup.connect(BASE_URL_BOOST).get();
+			Document doc = Jsoup.connect(BASE_URL_OPENJDK).get();
 			Elements tables = doc.select("table");
 			for (Element table : tables) {
-				Elements dateColumns = table.select(DATE_TAG_SELECTOR);
-				for (Element column : dateColumns) {
-					Elements links = column.select("a");
-					if (links.size() > 0) {
-						Element link = links.first();
-						if (link.text().endsWith("Date")) {
-							String url = BASE_URL_BOOST + link.attr("href");
-							dateUrls.add(url);
+				Elements tr_tags = table.select(TR_TAG_SELECTOR);
+				tr_tags.remove(0); // remove cabecalho da tabela
+				for (Element tag : tr_tags) {
+					Elements dateColumns = tag.select(DATE_TAG_SELECTOR);
+					for (Element column : dateColumns) {
+						Elements links = column.select("a");
+						if (links.size() > 0) {
+							for (Element link : links) {
+								if (link.text().endsWith("[ Date ]")) {
+									String url = BASE_URL_OPENJDK + link.attr("href");
+									logger.info(url); // debug
+									dateUrls.add(url);
+								}
+							}
 						}
 					}
 				}
@@ -94,14 +96,14 @@ public class ScrapBoost implements PageScraper {
 
 	/**
 	 * método é responsável por obter os links das threads a partir de um documento
-	 * HTML, a partir da BASE_URL_BOOST ou da url passada.
+	 * HTML, a partir da BASE_URL_OPENJDK ou da url passada.
 	 * 
 	 * @return Retorna a lista urls contendo as URLs das threads extraídas.
 	 */
 	public List<String> getLinksByThread() throws IOException {
 		List<String> threadUrls = new LinkedList<>();
 		try {
-			Document doc = Jsoup.connect(BASE_URL_BOOST).get();
+			Document doc = Jsoup.connect(BASE_URL_OPENJDK).get();
 			Elements tables = doc.select("table");
 			for (Element table : tables) {
 				Elements threadColumns = table.select(THREAD_TAG_SELECTOR);
@@ -110,7 +112,7 @@ public class ScrapBoost implements PageScraper {
 					if (links.size() > 0) {
 						Element link = links.first();
 						if (link.text().endsWith("Thread")) {
-							String url = BASE_URL_BOOST + link.attr("href");
+							String url = BASE_URL_OPENJDK + link.attr("href");
 							threadUrls.add(url);
 						}
 					}
@@ -132,15 +134,10 @@ public class ScrapBoost implements PageScraper {
 	public Elements extractLiTags(Document doc) {
 		Elements liTags = new Elements();
 		try {
-			Element ulParent = doc.select("ul").first();
+			Element ulParent = doc.select("ul").get(1);
 			Elements lis = ulParent.select("li");
 			for (Element li : lis) {
 				liTags.add(li);
-				/*
-				 * Element ul = li.select("ul").first(); if (ul != null) { Document newDoc =
-				 * Jsoup.parse(ul.toString()); // Chama recursivamente o método
-				 * liTags.addAll(extractLiTags(newDoc)); } // break;
-				 */
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -150,7 +147,7 @@ public class ScrapBoost implements PageScraper {
 
 	/**
 	 * Método é responsável por obter os links das mensagens a partir das URLs das
-	 * threads
+	 * threads organizadas por data.
 	 * 
 	 * @return Retorna um conjunto msgs contendo os links das mensagens extraídas.
 	 */
@@ -158,25 +155,30 @@ public class ScrapBoost implements PageScraper {
 	public Set<String> getLinksMessages() throws IOException {
 		Set<String> msgs = new HashSet<String>();
 		List<String> urls = getLinksByDate();
-		// List<String> urls = getLinksByThread();
 		for (String url : urls) {
 			try {
 				Connection connection = Jsoup.connect(url);
 				Document doc = connection.get();
 				for (Element li : extractLiTags(doc)) {
-					String link = li.select("a").attr("href");
+					String link = li.select("a").first().attr("href");
 					msgs.add(url.replace(DATE_URL_SUFFIX, link));
 				}
-//				break;
+				// break; // debug
 			} catch (Exception e) {
 				logger.error("Error while getting links messages for URL: " + url, e);
 				fileLogger.logException("Error while getting links messages for URL:", "url", e);
 			}
 		}
-//		logger.info(msgs.size());
+		logger.info(msgs.size()); // debug
 		return msgs;
 	}
 
+	/**
+	 * Método que percorre uma lista de URLs, faz a raspagem de dados dessas URLs
+	 * para criar objetos Post e armazena esses objetos em uma lista.
+	 * 
+	 * @return Retorna uma lista de objetos do tipo Post
+	 */
 	/**
 	 * Método que percorre uma lista de URLs, faz a raspagem de dados dessas URLs
 	 * para criar objetos Post e armazena esses objetos em uma lista.
@@ -215,29 +217,17 @@ public class ScrapBoost implements PageScraper {
 	 */
 	public void retrieveAuthorAndDate(Document doc, Post post) {
 		try {
-			Element headers = doc.select("p.headers").first();
-			String headersContent = headers.text();
+			Element author = doc.select("b").first();
+			String authorName = author.text();
 
-			String name = "";
-			String date = "";
+			Element date = doc.select("i").first();
+			String dateString = date.text().substring(4); // remove dia da semana
 
-			Pattern authorPattern = Pattern.compile("From:\\s(.+?)\\s\\(");
-			Matcher authorMatcher = authorPattern.matcher(headersContent);
-			if (authorMatcher.find()) {
-				name = authorMatcher.group(1).trim();
-			}
+			logger.info("Autor: " + authorName); // debug
+			logger.info("Data: " + dateString); // debug
 
-			Pattern datePattern = Pattern.compile("Date:\\s(\\d{4}-\\d{2}-\\d{2}\\s\\d{2}:\\d{2}:\\d{2})");
-			Matcher dateMatcher = datePattern.matcher(headersContent);
-			if (dateMatcher.find()) {
-				date = dateMatcher.group(1).trim();
-			}
-
-			logger.info("Autor: " + name);
-			logger.info("Data: " + date);
-
-			post.setAuthorName(utf8EncodedString(name));
-			post.setPublicationDate(date);
+			post.setAuthorName(utf8EncodedString(authorName));
+			post.setPublicationDate(dateString);
 		} catch (Exception e) {
 			logger.error("Error while retrieving author and date", e);
 			fileLogger.logException("Error while retrieving author and date", "url", e);
@@ -254,11 +244,10 @@ public class ScrapBoost implements PageScraper {
 	 */
 	public void retrieveTitle(Document doc, Post post) {
 		try {
-			Elements title = doc.select("title");
+			Elements title = doc.select("h1	");
 			for (Element t : title) {
 				String text = t.text();
-				text = text.replace("Boost mailing page: ", ""); // remove título supérfluo
-				logger.info("title: " + text);
+				logger.info("title: " + text); // debug
 				post.setTitle(utf8EncodedString(text));
 			}
 		} catch (Exception e) {
@@ -279,19 +268,10 @@ public class ScrapBoost implements PageScraper {
 	 */
 	public void retrieveBody(Document doc, Post post) {
 		try {
-
-			String start = "<!-- body=\"start\" -->";
-			String end = "<!-- body=\"end\" -->";
-			String tagBody = getStringBetweenTwoCharacters(doc.toString(), start, end);
-			doc = Jsoup.parse(tagBody);
-
-			Elements paragraphs = doc.select("p");
-			List<String> paragraphTexts = paragraphs.stream().map(Element::text).map(String::trim)
-					.filter(text -> !text.isEmpty()).collect(Collectors.toList());
-
-			String body = String.join("\n", paragraphTexts);
+			Elements textsection = doc.select("pre");
+			String body = textsection.text();
 			post.setBody(utf8EncodedString(body));
-			// System.out.println(post.getBody());
+			// System.out.println(post.getBody()); // debug
 		} catch (Exception e) {
 			logger.error("Error while retrieving body", e);
 			fileLogger.logException("Error while retrieving body", "url", e);
