@@ -1,6 +1,6 @@
 package br.unb.scrap.enums;
 
-public class PostTypeTest {
+public class PostTypeEnumTest {
 
 	public static void main(String[] args) {
 		testOriginalPostType();
@@ -8,13 +8,13 @@ public class PostTypeTest {
 	}
 
 	private static void testOriginalPostType() {
-		PostType originalType = PostType.ORIGINAL;
+		PostTypeEnum originalType = PostTypeEnum.ORIGINAL;
 		boolean value = originalType.isValue();
 		System.out.println("This is the original email" + " >>>>>>" + value);
 	}
 
 	private static void testReplyPostType() {
-		PostType replyType = PostType.REPLY;
+		PostTypeEnum replyType = PostTypeEnum.REPLY;
 		boolean value = replyType.isValue();
 		System.out.println("This is the reply email" + " >>>>>>" + value);
 	}
