@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import br.unb.scrap.enums.PostType;
+import br.unb.scrap.enums.PostTypeEnum;
 
 public class PostTest {
 
@@ -52,9 +52,9 @@ public class PostTest {
 
 	@Test
 	public void testSetAndGetPostType() {
-		PostType postType = PostType.ORIGINAL;
+		PostTypeEnum postType = PostTypeEnum.ORIGINAL;
 		post.setPostType(postType);
-		Assertions.assertEquals(postType, post.getPostType(), "Message type not set correctly.");
+		Assertions.assertEquals(postType, post.getPostTypeEnum(), "Message type not set correctly.");
 	}
 
 	@Test
@@ -64,7 +64,7 @@ public class PostTest {
 		String date = "04-05-1998";
 		String subject = "Software Rejuvenation";
 		String body = "it is a phenomenon that leads to burnout...software system failure";
-		PostType postType = PostType.ORIGINAL;
+		PostTypeEnum postType = PostTypeEnum.ORIGINAL;
 
 		post.setId(id);
 		post.setAuthorName(name);
